blog: create uri rewrite log group explicitly
This required some trickery to create, namely, the original log group
needed to be deleted first and then, quickly, the changeset needed to be
applied.

Interestingly enough, if this is done quickly enough, the original log
data is not actually lost, as of this writing.  I personally don't
think that was intended, however, I think I can understand how something
like this could happen.

It would be nice if it were possible to get the name of the log group
that will be used from the lambda function resource.  However, this does
not currently seem possible.

Signed-off-by: Kenny Ballou <kballou@devnulllabs.io>